requirements.txt
pdfplumber==0.11.4
python-dateutil==2.9.0.post0
pandas==2.2.2
PyYAML==6.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
PyMuPDF==1.23.26
Pillow==10.0.1
numpy<2.0.0

